ui package - qrcode cleanup
