+++ packages/renderer/package.json
@@ -56,7 +56,6 @@
     "date-fns": "^3.6.0",
     "framer-motion": "^11.1.1",
     "html5-qrcode": "^2.3.8",
-    "qrcode-generator": "^1.4.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-error-boundary": "^4.0.12",
