+++ packages/ui/package.json
@@ -44,6 +44,8 @@
     "@w3ux/utils": "0.4.0-alpha.0",
     "bignumber.js": "^9.1.2",
     "framer-motion": "^11.1.1",
+    "html5-qrcode": "^2.3.8",
+    "qrcode-generator": "^1.4.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "styled-components": "^6.1.8"
